requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
pysrt==1.1.2
tenacity==8.0.1
pytest==6.2.5